TextBlock Final/TB/src/TextBlock.java
/**
 * Simple blocks of text for a lab on polymorphism
 * 
 * @author Samuel A. Rebelsky & Luke Walters
 * @version 1.2 of February 2019
 */
public interface TextBlock {
  // +---------+-----------------------------------------------------------
  // | Methods |
  // +---------+

  /**
   * Get one row from the block.
   * 
   * @pre i < this.height()
   * @exception Exception if the row number is invalid.
   */
  public String row(int i) throws Exception;

  /**
   * Determine how many rows are in the block.
   */
  public int height();

  /**
   * Determine how many columns are in the block.
   */
  public int width();

  /**
   * Equality method that checks to see if a two TextBlocks were constructed in
   * the same manner
   * 
   * @param tbCmp
   * @return
   */
  public boolean eqv(TextBlock tbCmp);

} // interface TextBlock
